Separate entity class from api
