+++ lib/tracker.rb
@@ -1,5 +1,6 @@
-require "tracker/version"
+require 'tracker/version'
 require 'tracker/api'
+require 'tracker/entities'
 
 module Tracker
   # Your code goes here...
